+++ src/main/java/org/mmarini/wheelly/envs/RobotEnv.java
@@ -367,6 +367,7 @@ public class RobotEnv implements Environment {
         distance = Nd4j.createFromArray((float) status.getEchoDistance());
         canMoveForward = Nd4j.createFromArray(status.canMoveForward() ? 1F : 0F);
         canMoveBackward = Nd4j.createFromArray(status.canMoveBackward() ? 1F : 0F);
-        contacts = Nd4j.createFromArray((float) status.getContacts());
+        // TODO fix contacts
+        contacts = Nd4j.createFromArray(status.isFrontSensors() ? 1F : 0F, status.isRearSensors() ? 1F : 0F);
     }
 }