+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Fixed
 
 - Issue #280: RobotExecutor stuck
+- Issue #284: Wrong speed with simulated robot
 
 ## [0.11.1] 2023-12-24